[New] `focusTracker.ts`
